+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 otj-kafka changelog
 ===================
 
+2.3.0
+-----
+
+* added offset monitoring utilities
+* embedded kafka broker will now ready itself
+
 2.2.0
 -----